fastrand

Function choice

pub fn choice<I>(iter: I) -> Option<I::Item>
where
    I: IntoIterator,
    I::IntoIter: ExactSizeIterator,

Available on crate feature std only.

Choose an item from an iterator at random.

This function may have an unexpected result if the len() property of the iterator does not match the actual number of items in the iterator. If the iterator is empty, this returns None.