Struct fastrand::Rng

pub struct Rng(_);

A random number generator.

Implementations

impl Rng

pub fn new() -> Rng

Creates a new random number generator.

pub fn with_seed(seed: u64) -> Self

Creates a new random number generator with the initial seed.

pub fn alphabetic(&self) -> char

Generates a random char in ranges a-z and A-Z.

pub fn alphanumeric(&self) -> char

Generates a random char in ranges a-z, A-Z and 0-9.

pub fn bool(&self) -> bool

Generates a random bool.

pub fn digit(&self, base: u32) -> char

Generates a random digit in the given base.

Digits are represented by chars in ranges 0-9 and a-z.

Panics if the base is zero or greater than 36.

pub fn f32(&self) -> f32

Generates a random f32 in range 0..1.

pub fn f64(&self) -> f64

Generates a random f64 in range 0..1.

pub fn i8(&self, range: impl RangeBounds<i8>) -> i8

Generates a random i8 in the given range.

Panics if the range is empty.

pub fn i16(&self, range: impl RangeBounds<i16>) -> i16

Generates a random i16 in the given range.

Panics if the range is empty.

pub fn i32(&self, range: impl RangeBounds<i32>) -> i32

Generates a random i32 in the given range.

Panics if the range is empty.

pub fn i64(&self, range: impl RangeBounds<i64>) -> i64

Generates a random i64 in the given range.

Panics if the range is empty.

pub fn i128(&self, range: impl RangeBounds<i128>) -> i128

Generates a random i128 in the given range.

Panics if the range is empty.

pub fn isize(&self, range: impl RangeBounds<isize>) -> isize

Generates a random isize in the given range.

Panics if the range is empty.

pub fn lowercase(&self) -> char

Generates a random char in range a-z.

pub fn seed(&self, seed: u64)

Initializes this generator with the given seed.

pub fn get_seed(&self) -> u64

Gives back current seed that is being held by this generator.

pub fn shuffle<T>(&self, slice: &mut [T])

Shuffles a slice randomly.

pub fn fill(&self, slice: &mut [u8])

Fill a byte slice with random data.

pub fn u8(&self, range: impl RangeBounds<u8>) -> u8

Generates a random u8 in the given range.

Panics if the range is empty.

pub fn u16(&self, range: impl RangeBounds<u16>) -> u16

Generates a random u16 in the given range.

Panics if the range is empty.

pub fn u32(&self, range: impl RangeBounds<u32>) -> u32

Generates a random u32 in the given range.

Panics if the range is empty.

pub fn u64(&self, range: impl RangeBounds<u64>) -> u64

Generates a random u64 in the given range.

Panics if the range is empty.

pub fn u128(&self, range: impl RangeBounds<u128>) -> u128

Generates a random u128 in the given range.

Panics if the range is empty.

pub fn usize(&self, range: impl RangeBounds<usize>) -> usize

Generates a random usize in the given range.

Panics if the range is empty.

pub fn uppercase(&self) -> char

Generates a random char in range A-Z.

pub fn char(&self, range: impl RangeBounds<char>) -> char

Generates a random char in the given range.

Panics if the range is empty.