Module fastrand::global_rng

A global, thread-local random number generator.

Macros

integer ğŸ”’

Structs

RestoreOnDrop ğŸ”’

Make sure the original RNG is restored even on panic.

Constants

DEFAULT_RNG_SEED ğŸ”’

RNG ğŸ”’

Functions

alphabetic

Generates a random char in ranges a-z and A-Z.

alphanumeric

Generates a random char in ranges a-z, A-Z and 0-9.

bool

Generates a random bool.

char

Generates a random char in the given range.

choice

Choose an item from an iterator at random.

choose_multiple

Collects amount values at random from the iterator into a vector.

digit

Generates a random digit in the given base.

f32

Generates a random f32 in range 0..1.

f64

Generates a random f64 in range 0..1.

get_seed

Gives back current seed that is being held by the thread-local generator.

i8

Generates a random i8 in the given range.

i16

Generates a random i16 in the given range.

i32

Generates a random i32 in the given range.

i64

Generates a random i64 in the given range.

i128

Generates a random i128 in the given range.

isize

Generates a random isize in the given range.

lowercase

Generates a random char in range a-z.

random_seed ğŸ”’

seed

Initializes the thread-local generator with the given seed.

shuffle

Shuffles a slice randomly.

try_with_rng ğŸ”’

Try to run an operation with the current thread-local generator.

u8

Generates a random u8 in the given range.

u16

Generates a random u16 in the given range.

u32

Generates a random u32 in the given range.

u64

Generates a random u64 in the given range.

u128

Generates a random u128 in the given range.

uppercase

Generates a random char in range A-Z.

usize

Generates a random usize in the given range.

with_rng ğŸ”’

Run an operation with the current thread-local generator.