fastrand

Struct Rng

pub struct Rng(/* private fields */);

A random number generator.

Implementations

impl Rng

pub fn new() -> Rng

Creates a new random number generator.

impl Rng

pub fn with_seed(seed: u64) -> Self

Creates a new random number generator with the initial seed.

pub fn fork(&mut self) -> Self

Clones the generator by deterministically deriving a new generator based on the initial seed.

This function can be used to create a new generator that is a โspinoffโ of the old generator. The new generator will not produce the same sequence of values as the old generator.

Example

// Seed two generators equally, and clone both of them.
let mut base1 = fastrand::Rng::with_seed(0x4d595df4d0f33173);
base1.bool(); // Use the generator once.

let mut base2 = fastrand::Rng::with_seed(0x4d595df4d0f33173);
base2.bool(); // Use the generator once.

let mut rng1 = base1.fork();
let mut rng2 = base2.fork();

println!("rng1 returns {}", rng1.u32(..));
println!("rng2 returns {}", rng2.u32(..));

pub fn alphabetic(&mut self) -> char

Generates a random char in ranges a-z and A-Z.

pub fn alphanumeric(&mut self) -> char

Generates a random char in ranges a-z, A-Z and 0-9.

pub fn bool(&mut self) -> bool

Generates a random bool.

pub fn digit(&mut self, base: u32) -> char

Generates a random digit in the given base.

Digits are represented by chars in ranges 0-9 and a-z.

Panics if the base is zero or greater than 36.

pub fn f32(&mut self) -> f32

Generates a random f32 in range 0..1.

pub fn f64(&mut self) -> f64

Generates a random f64 in range 0..1.

pub fn choose_multiple<I: IntoIterator>( &mut self, source: I, amount: usize, ) -> Vec<I::Item>

Collects amount values at random from the iterable into a vector.

The length of the returned vector equals amount unless the iterable contains insufficient elements, in which case it equals the number of elements available.

Complexity is O(n) where n is the length of the iterable.

pub fn i8(&mut self, range: impl RangeBounds<i8>) -> i8

Generates a random i8 in the given range.

Panics if the range is empty.

pub fn i16(&mut self, range: impl RangeBounds<i16>) -> i16

Generates a random i16 in the given range.

Panics if the range is empty.

pub fn i32(&mut self, range: impl RangeBounds<i32>) -> i32

Generates a random i32 in the given range.

Panics if the range is empty.

pub fn i64(&mut self, range: impl RangeBounds<i64>) -> i64

Generates a random i64 in the given range.

Panics if the range is empty.

pub fn i128(&mut self, range: impl RangeBounds<i128>) -> i128

Generates a random i128 in the given range.

Panics if the range is empty.

pub fn isize(&mut self, range: impl RangeBounds<isize>) -> isize

Generates a random isize in the given range.

Panics if the range is empty.

pub fn lowercase(&mut self) -> char

Generates a random char in range a-z.

pub fn seed(&mut self, seed: u64)

Initializes this generator with the given seed.

pub fn get_seed(&self) -> u64

Gives back current seed that is being held by this generator.

pub fn choice<I>(&mut self, iter: I) -> Option<I::Item>
where I: IntoIterator, I::IntoIter: ExactSizeIterator,

Choose an item from an iterator at random.

This function may have an unexpected result if the len() property of the iterator does not match the actual number of items in the iterator. If the iterator is empty, this returns None.

pub fn shuffle<T>(&mut self, slice: &mut [T])

Shuffles a slice randomly.

pub fn fill(&mut self, slice: &mut [u8])

Fill a byte slice with random data.

pub fn u8(&mut self, range: impl RangeBounds<u8>) -> u8

Generates a random u8 in the given range.

Panics if the range is empty.

pub fn u16(&mut self, range: impl RangeBounds<u16>) -> u16

Generates a random u16 in the given range.

Panics if the range is empty.

pub fn u32(&mut self, range: impl RangeBounds<u32>) -> u32

Generates a random u32 in the given range.

Panics if the range is empty.

pub fn u64(&mut self, range: impl RangeBounds<u64>) -> u64

Generates a random u64 in the given range.

Panics if the range is empty.

pub fn u128(&mut self, range: impl RangeBounds<u128>) -> u128

Generates a random u128 in the given range.

Panics if the range is empty.

pub fn usize(&mut self, range: impl RangeBounds<usize>) -> usize

Generates a random usize in the given range.

Panics if the range is empty.

pub fn uppercase(&mut self) -> char

Generates a random char in range A-Z.

pub fn char(&mut self, range: impl RangeBounds<char>) -> char

Generates a random char in the given range.

Panics if the range is empty.